Add SiteTrip functionality and images to application
This commit includes the addition of a new SiteTrip functionality along with images. It includes the creation of models, views and templates for SiteTrip. It also adds urls for SiteTrip in the admin panel. Modifications were done in the application model to include a motivation text field. The front-end was also updated to support this new functionality. Images can be added to each SiteTrip, with the ability to view or delete them in the admin panel. The commit further involves deletion of two handler methods to update and delete applications in the api handler.